Implement aggregation + BVH system.
Added an abstract aggregate trait that allows organizing shapes in
different structures for different performance characteristics.
Also added a BVH implementation of aggregate, which uses the
Surface Area Heuristic for construction. I'm currently not seeing
the performance gains I expected, but the first pass at the tree
seems to have been effective in building the structure necessary,
so I'm pushing this now and planning to do an optimization pass later.